requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
python-dotenv==1.0.0
click==8.1.7
sqlalchemy==2.0.23
lxml==4.9.3
fake-useragent==1.4.0
